Cleaned up \Widget\ and added Layout
