+++ SS31.Client.UI/Widget/WidgetContainer.cs
@@ -6,6 +6,22 @@ using Microsoft.Xna.Framework.Graphics;
 
 namespace SS31.Client.UI
 {
+	// Interface for widgets that contain other widgets
+	public interface IWidgetContainer : IWidget
+	{
+		// The children of this widget
+		List<Widget> Children { get; }
+		// If this widget should use the stencil buffer to mask its children
+		bool MaskChildren { get; }
+
+		// Add a widget as a child
+		void AddChild(Widget w);
+		// Remove a widget instance as a child
+		void RemoveChild(Widget w);
+		// Remove a widget with the specified identifier as a child
+		void RemoveChild(string ident);
+	}
+
 	// A base class for widgets that contain other widgets, with the default IWidgetContainer implementation
 	public class WidgetContainer : Widget, IWidgetContainer
 	{
